+ Added working PPU for some Mapper0 games + Added debug windows for PPU (chr, nametables, sprites, palettes) + Added base implementation to support input + Added DMA transfer support + Added CPU interruptions handling + Added cycle accurate mode * Fixed read concatenation issues in release (non fixed order of execution) * Fixed stack increment/decrement issues * Fixed some CPU logging issues * Fixed BIT flags * Fixed Compare flags * Fixed mapper0 read

+++ Window.h
@@ -34,7 +34,7 @@ protected:
     int xPosition;
     int yPosition;
 
-    sf::Uint8* screenArray = nullptr;
+    sf::Uint32* screenArray = nullptr;
     sf::Texture screenTexture;
     sf::Sprite screenSprite;
